third_party/WebKit/LayoutTests/http/tests/inspector/network/network-choose-preview-view-expected.txt

Tests to make sure the proper view is used for the data that is received in network panel.


Running: testSimpleJSON
Creating a NetworkRequest with mimeType: application/json
Content: [533,3223]
Its previewer type: SearchableView > widget vbox json-view

Running: testMimeJSON
Creating a NetworkRequest with mimeType: application/vnd.document+json
Content: {foo0foo: 123}
Its previewer type: SearchableView > widget vbox json-view

Running: testSimpleXML
Creating a NetworkRequest with mimeType: text/xml
Content: <bar><foo/></bar>
Its previewer type: widget shadow-xml-view source-code

Running: testXMLMimeButJSON
Creating a NetworkRequest with mimeType: text/xml
Content: {foo0: 'barr', 'barr': 'fooo'}
Its previewer type: SearchableView > widget vbox json-view

Running: testXMLWithUnknownMime
Creating a NetworkRequest with mimeType: text/foobar
Content: <bar><foo/></bar>
Its previewer type: SearchableView > widget vbox

Running: testXMLWithError500
Creating a NetworkRequest with mimeType: text/xml
Content: <bar><foo/></bar>
Its previewer type: widget shadow-xml-view source-code

Running: testUnknownMimeTextWithError500
Creating a NetworkRequest with mimeType: text/foobar
Content: Foo Bar
Its previewer type: SearchableView > widget vbox

Running: testBinaryImageFile
Creating a NetworkRequest with mimeType: image/png
Content: Bin**NULL**ary File**NULL****NULL**
Its previewer type: SearchableView > widget vbox

Running: testBinaryBlankImageFile
Creating a NetworkRequest with mimeType: image/png
Content: 
Its previewer type: widget vbox empty-view
